made test of ClientesServiceImpl and begin HabitacionesServiceImpl test

+++ Proyecto-Hotel/src/main/java/com/proyectoHotel/controller/dto/HabitacionesDTO.java
@@ -26,7 +26,7 @@ public class HabitacionesDTO {
 
     private LocalDateTime fechaDeEgreso;
 
-    private BigDecimal precio;
+    private double precio;
 
     private Clientes clientes;
 }
