+++ Proyecto-Hotel/src/main/java/com/proyectoHotel/model/Habitaciones.java
@@ -5,9 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Getter
 @Setter
@@ -37,7 +35,7 @@ public class Habitaciones {
     private LocalDateTime fechaDeEgreso;
 
     @Column
-    private BigDecimal precio;
+    private double precio;
 
     @ManyToOne
     @JsonIgnore
